pyproject.toml
[project]
name = "matera"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "ruff>=0.14.10",
]
